+++ fud_demo/src/main/java/com/centerm/fud_demo/entity/Role.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class Role {
-    private Integer role_id;
+    private Long role_id;
     private String role_name;
 }
